simple vmm structures; ds/list.h; sthoth

+++ include/string.h
@@ -4,8 +4,8 @@
 // Get size_t, NULL from stddef.h
 #include <stddef.h>
 
-int strlen(const char *s);
-int strnlen(const char *s, size_t size);
+size_t strlen(const char *s);
+size_t strnlen(const char *s, size_t size);
 char *strcpy(char *dst, const char *src);
 char *strcat(char *dst, const char *src);
 char *strncpy(char *dst, const char *src, size_t size);
